+++ src/test/java/bio/terra/cloudres/google/bigquery/TabletCowTest.java
@@ -17,7 +17,9 @@ public class TabletCowTest {
 
   private static DatasetInfo reusableDataset;
 
-  private List<TableId> createdTableIds;
+  // Cleanup the createdTableIds list and use this to track tables created in one test method.
+  // In this way to can make sure tables can always be cleaned up even test failed in the middle.
+  private List<TableId> createdTableIds = new ArrayList<>();
 
   @BeforeAll
   public static void createReusableDataset() {
@@ -28,13 +30,6 @@ public class TabletCowTest {
   @AfterAll
   public static void deleteReusableDataset() {
     bigQueryCow.delete(REUSABLE_DATASET_ID);
-  }
-
-  @BeforeEach
-  public void setUp() {
-    // Cleanup the createdTableIds list and use this to track tables created in one test method.
-    // In this way to can make sure tables can always be cleaned up even test failed in the middle.
-    createdTableIds = new ArrayList<>();
   }
 
   @AfterEach
